src/requirements.txt
django==2.2.2
psycopg2-binary==2.8.5
djangorestframework==3.11.0
djoser==2.0.3
gevent==1.4.0
eventlet==0.25.1
ics==0.7
ipython==7.13.0
requests==2.21.0
redis==3.3.11
Pygments==2.5.2
kombu==4.6.6
Cython==0.29.14
celery==4.3.0